+++ test/test_distributed.py
@@ -20,6 +20,7 @@ import torch.nn.functional as F
 from common_utils import TestCase, run_tests, skipIfRocm
 from torch._utils_internal import TEST_MASTER_ADDR as MASTER_ADDR
 from torch._utils_internal import TEST_MASTER_PORT as MASTER_PORT
+from common_distributed import simple_sparse_reduce_tests
 
 try:
     import torchvision
@@ -1012,6 +1013,29 @@ class _DistTestBase(object):
         self._test_all_reduce_helper(
             group, group_id, rank, dist.ReduceOp.MAX, -1, 10, 10
         )
+
+    # SPARSE ALL REDUCE
+    def _test_sparse_all_reduce_sum(self, fn):
+        group, group_id, rank = self._init_global_test()
+
+        tests = simple_sparse_reduce_tests(
+            rank,
+            dist.get_world_size(),
+            num_inputs=1)
+        for (inputs, outputs) in tests:
+            tensors = [fn(input) for input in inputs]
+            dist.all_reduce(tensors[0], dist.ReduceOp.SUM, group_id)
+            self.assertEqual(tensors[0], outputs[0])
+
+    @unittest.skipIf(BACKEND == "nccl", "Nccl does not support all reduce on sparse tensors")
+    def test_sparse_all_reduce_sum(self):
+        self._test_sparse_all_reduce_sum(lambda t: t)
+
+    @unittest.skipIf(BACKEND == "nccl", "Nccl does not support all reduce on sparse tensors")
+    @skip_if_not_multigpu
+    @skip_if_rocm
+    def test_sparse_all_reduce_sum_cuda(self):
+        self._test_sparse_all_reduce_sum(lambda t: t.clone().cuda())
 
     # ALL REDUCE - COALESCED
     @staticmethod
