+++ torch/distributed/distributed_c10d.py
@@ -908,7 +908,6 @@ def all_reduce(tensor,
     else:
         work.wait()
 
-
 def all_reduce_coalesced(tensors,
                          op=ReduceOp.SUM,
                          group=group.WORLD,
